feat(phi-plugin): 移植 `phisong`
- 更新同步版本至 9a34bf4
- 添加插件功能测试相关说明
- 完成移植 `phisong`
- 优化帮助信息处理逻辑
- 简化 manage 模块中的代码
- 当前所有模块已移植完成，正在进行可用性测试

+++ phi-plugin/model/getdata.py
@@ -377,7 +377,7 @@ class getdata:
         return getInfo.SongGetId(song)
 
     @staticmethod
-    def getrks(acc: float, difficulty: int):
+    def getrks(acc: float, difficulty: float):
         """
         计算等效rks
 
@@ -396,7 +396,7 @@ class getdata:
             return difficulty * (((acc - 55) / 45) ** 2)
 
     @staticmethod
-    def comsuggest(rks: float, difficulty: int, count: int | None = None):
+    def comsuggest(rks: float, difficulty: float, count: int | None = None):
         """
         计算所需acc
